[ft-increase-incidents-controller-test-160445645] refactor code to reduce complexity and make the code more readable
