+++ src/server/controllers/incidents.js
@@ -1,57 +1,13 @@
 const errorLogs = require('./errorLogs');
 const Incident = require('../models').Incidents;
 const User = require('../models').Users;
-const Location = require('../models').Locations;
-const Level = require('../models').Levels;
-const Status = require('../models').Statuses;
 const AssigneeModel = require('../models').assigneeIncidents;
 const LocationService = require('../controllers/locations');
 const findOrCreateUser = require('../helpers/findOrCreateUser');
 const listAssigneeIncidentsIncludes = require('../helpers/listAssigneeIncidentsIncludes');
-const generateAssigneeOrCcdEmailBody = require('../helpers/generateAssigneeOrCcdEmailBody');
-const emailHelper = require('../helpers/emailHelper');
+const { addAssignee, addCcdUser, findIncidentById, returnIncidentsIncludes } = require('../helpers/incidentHelper');
 
-let userAttributes = ['username', 'slackId', 'imageUrl', 'email'];
-
-let includes = [
-  {
-    model: Level,
-    attributes: ['name']
-  },
-  {
-    model: Status,
-    attributes: ['status']
-  },
-  {
-    model: Location,
-    attributes: ['name', 'centre', 'country']
-  },
-  {
-    model: User,
-    as: 'assignees',
-    userAttributes,
-    through: {
-      attributes: ['assignedRole']
-    }
-  },
-  {
-    model: User,
-    as: 'reporter',
-    userAttributes,
-    through: {
-      attributes: []
-    }
-  },
-  {
-    model: User,
-    as: 'witnesses',
-    userAttributes,
-    through: {
-      attributes: []
-    }
-  }
-];
-
+const include = returnIncidentsIncludes();
 // mapping Assignees
 const mapAssignees = incident => {
   return incident.map(oneIncident => {
@@ -59,56 +15,6 @@ const mapAssignees = incident => {
       oneIncident.dataValues.assigneeIncidents.assignedRole;
     delete oneIncident.dataValues.assigneeIncidents;
     return oneIncident;
-  });
-};
-
-const findIncidentById = (id, res) => {
-  return Incident.findById(id, { include: includes })
-    .then(incident => {
-      return incident;
-    })
-    .catch(error => {
-      throw error;
-    });
-};
-
-/**
- * @function getUserDetails
- * @param payload object
- * @return userDetails object
- */
-const getUserDetails = async payload => {
-  let userDetails;
-  if (Array.isArray(payload)) {
-    payload.map(async ccdUser => {
-      userDetails = await User.findById(ccdUser.userId);
-      userDetails.dataValues.incidentId = ccdUser.incidentId;
-      return userDetails;
-    });
-    return userDetails;
-  }
-  userDetails = await User.findById(payload.userId);
-  userDetails.dataValues.incidentId = payload.incidentId;
-  return userDetails;
-};
-
-/**
- * @function sendAssigneeOrCcdEmail
- * @param payload object
- * @return error or success message
- */
-
-const sendAssigneeOrCcdEmail = async payload => {
-  const userDetails = await getUserDetails(payload);
-  const emailBody = await generateAssigneeOrCcdEmailBody({
-    ...userDetails.dataValues,
-    assignedRole: payload.assignedRole // add assigned role
-  });
-  emailHelper.sendMail(emailBody, error => {
-    if (error) {
-      return error;
-    }
-    return { message: 'The email was sent successfully' };
   });
 };
 
@@ -224,7 +130,7 @@ module.exports = {
         .send({ data: { incidents: mappedIncidents }, status: 'success' });
     }
     return Incident.findAll({
-      include: includes
+      include
     })
       .then(incidents => {
         let mappedIncidents = incidents.map(incident => {
@@ -263,44 +169,24 @@ module.exports = {
 
   // update an incident
   update(req, res) {
-    let assignedUser = req.body.assignee;
-    let ccdUser = req.body.ccd;
-    let destroyCcdPromise;
-    let addCcdPromises = [];
+    let { assignee: assignedUser, ccd: ccdUser } = req.body; 
 
     let findIncidentPromise = Incident.findById(req.params.id, {
-      include: includes
+      include
     }).then(incident => {
       return incident
         ? Promise.resolve(incident)
         : Promise.reject({
-            message: 'Incident not found',
-            status: 'fail'
-          });
+          message: 'Incident not found',
+          status: 'fail'
+        });
     });
 
     if (assignedUser) {
       return findIncidentPromise
         .then(incident => {
           if (incident.dataValues.assignees.length === 0) {
-            return (
-              User.findById(assignedUser.userId)
-                // new assignee
-                .then(async assignee => {
-                  assignedUser.assignedRole = 'assignee';
-                  await sendAssigneeOrCcdEmail(assignedUser);
-                  assignee.assigneeIncidents = {
-                    assignedRole: 'assignee'
-                  };
-                  return incident.addAssignee(assignee);
-                })
-                .then(() => {
-                  return findIncidentById(incident.id, res);
-                })
-                .then(data => {
-                  return res.status(200).send({ data, status: 'success' });
-                })
-            );
+            return addAssignee({ assignedUser, incident, res });
           } else {
             return (
               AssigneeModel.destroy({
@@ -310,23 +196,12 @@ module.exports = {
                 }
               })
                 .then(() => {
-                  return User.findById(assignedUser.userId);
-                })
-                // replacing an assignee
-                .then(async assignee => {
-                  assignedUser.assignedRole = 'assignee';
-                  await sendAssigneeOrCcdEmail(assignedUser);
-                  assignee.assigneeIncidents = {
-                    assignedRole: 'assignee'
-                  };
-                  return incident.addAssignee(assignee);
-                })
-                .then(() => {
-                  return findIncidentById(incident.id, res);
-                })
-                .then(data => {
-                  return res.status(200).send({ data, status: 'success' });
-                })
+                  return addAssignee({
+                    assignedUser,
+                    incident,
+                    res
+                  });
+                })                
             );
           }
         })
@@ -338,22 +213,7 @@ module.exports = {
       return findIncidentPromise
         .then(incident => {
           if (incident.dataValues.assignees.length === 0) {
-            const ccdPromises = ccdUser.map(async user => {
-              const ccd = User.findById(user.userId);
-              let currentCcd = { ...user, assignedRole: 'ccd' };
-              await sendAssigneeOrCcdEmail(currentCcd);
-              ccd.assigneeIncidents = {
-                assignedRole: 'ccd'
-              };
-              return incident.addAssignee(ccd);
-            });
-            return Promise.all(ccdPromises)
-              .then(() => {
-                return findIncidentById(incident.id, res);
-              })
-              .then(data => {
-                return res.status(200).send({ data, status: 'success' });
-              });
+            return addCcdUser({ ccdUser, res, incident });
           } else {
             return AssigneeModel.destroy({
               where: {
@@ -362,27 +222,11 @@ module.exports = {
               }
             })
               .then(() => {
-                const ccdPromises = ccdUser.map(async user => {
-                  const ccd = User.findById(user.userId);
-                  let currentCcd = { ...user, assignedRole: 'ccd' };
-                  await sendAssigneeOrCcdEmail(currentCcd);
-                  ccd.assigneeIncidents = {
-                    assignedRole: 'ccd'
-                  };
-                  return incident.addAssignee(ccd);
-                });
-                return Promise.all(ccdPromises);
-              })
-              .then(() => {
-                return findIncidentById(incident.id, res);
-              })
-              .then(data => {
-                return res.status(200).send({ data, status: 'success' });
+                return addCcdUser({ ccdUser, res, incident });
               });
           }
         })
         .catch(error => {
-          console.log({ error });
           errorLogs.catchErrors(error);
           return res.status(400).send(error);
         });
@@ -434,7 +278,7 @@ module.exports = {
       return res.status(400).send({ message: 'please provide query' });
     }
     return Incident.findAll({
-      include: includes,
+      include,
       where: {
         $or: [
           { subject: { $ilike: `%${req.query.q}%` } },
@@ -459,7 +303,7 @@ module.exports = {
       where: {
         categoryId: req.params.id
       },
-      include: includes
+      include
     })
       .then(incident => {
         res
